Removal of meta tags because of obsolecense
Meta tags are no longer used by major search engines to index webpages
and thus it is unnecessary to have them available in HTML shell.  I
didn't change the  readme.txt file but for some reasion it is showing up
in uncommitted changes and I can't stop it from doing so.